refactors : added modules; rename scripnt.js -> index.js; Datepicker, date.js hes been rewritten as pure functions
